+++ source/extensions/lightspeed.trex.properties_pane.widget/docs/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.15.3]
+### Changed
+- Reverted back to Label identifier for item_prim with new NicknameField
 
 ## [1.15.2]
 ### Changed